Teach the bot how to push boulders.
The bot will now push boulders before it tries to search everything.

If the boulder doesn't move, it marks the boulder as not being pushable from a certain direction.

If the boulder doesn't move because there is a monster behind it, the boulder is temporarily marked
as not pushable from a certain direction.